+++ packages/vue-fabric-fiber/lib/shapes/types.ts
@@ -1,3 +1,7 @@
 export type FabricObjectModel<T, ExtraOmit extends keyof T = never> = Partial<
   Omit<T, 'clipPath' | 'canvas' | ExtraOmit>
->
+> & {
+  leftPercent?: number
+  topPercent?: number
+  [key: string]: unknown
+}